Implement POC: Background NFC tag handling
The NFC Tag should be handled in the background. Unfortunately, NFC tags exclusively start activities. It's not possible to start a service or broadcast receiver.

This is why we have a dummy activity NfcActivity that simply starts NfcWorker with the WorkManager.

The NfcWorker then does the magic in the background.

Also rename the package and app id from de.eucalypto to net.eucalypto

#3