Add button to save category during edit
The previous behavior saved the category in onPause(), which occurs often: for example during screen rotation. The upside is that the user can navigate back from this fragment and it will be saved.
But that's not necessarily what the user expects.

Now it's clear that the button will save the category and nothing else. By clicking this button the app navigates back to the list of categories.

#8