Feature: Edit existing category
Now the user can edit an existing category. That is, change the name. The UUID is not changeable so that we don't have to update all entries in the database.

We refactored EditCategoryFragment a bit: now it receives an existing Category object as NavArg argument and doesn't care where it comes from. It simply displays the name of this category and writes/updates it in the database when done button is clicked.

If you want to edit an existing category, simply give this category to the navigation component's action as argument.

If you want to create a new category, then give a new (default values) category to this fragment.

The fragment doesn't care either way. :)

#8