Fix navigation bug
The previous code had flaws in navigation. For instance I defined a separate action from the edit fragment back to the category list fragment. But instead, I should have popped the navigation stack. Because once back in the category list fragment, when pressing back button, it tries to navigate "back" to the edit fragment, which is no longer there. Instead it should close the app. So the navigation stack got messed up.

The solution was to use either navController.popBackStack() or the current solution requireActivity().onBackPressed().