Refactor: Introduce extra Parcelable class for Category
I want to have my domain model classes as clean (pojo) as possible. (This is why there is an extra database representation.) And now there is a parcelable implementation with conversion functions that the client code has to use (that is the downside of this).

This introduces more classes but it allows for clear boundaries and separation of concern.

#6