Make write NFC tag activity look like a dialog
This is way more convenient. Writing the NFC tag should have been a dialog from the start, but it needs to be a full activity to be able to intercept NFC intents. Luckily, the theme for this activity is now set to a dialog in the manifest. Now it looks and feels like a dialog and can be closed by simply clicking outside it on the darkened background screen.

#4