Set up action bar with navigation component
This task itself is quite straightforward: call the extension function setupActionBarWithNavController(navController).

The tricky part for this code is getting the NavController in onCreate. The standard way of findNavController(R.id.nav_host_fragment) throws a RuntimeException.
So we have to actually "find" the navController in a more elaborate way, which is officially recommended:
https://developer.android.com/guide/navigation/navigation-getting-started#navigate
see also:
https://stackoverflow.com/questions/58703451/fragmentcontainerview-as-navhostfragment

We have to use this elaborate way because in main_activity.xml we're using FragmentContainerView as the NavHostFragment and not simply <fragment/>.

#13