Use navigation component in the overflow menu to navigate to Category List fragment
We still have to inflate a menu but the menu item id's have to match exactly id's of destinations or actions in the navigation graph.

We also added some labels in the navigation graph for the destinations. These labels are displayed in the action bar as titles because we set it up to work with navigation.

#13