Use navigation drawer to navigate to category list
Because navigation drawers are cool! Also, I will probably add a couple of more destinations that the user might want to quickly navigate to: a data visualizing screen (maybe even several), and a data export screen?

The overflow menu is now displayed in the whole main activity and leads to less frequently viewed pages like about and legal.

#13