Change model data: start/end time class
We previously used LocalDateTime which has some drawbacks: it is NOT a moment (instant) in time because it has no timezone data. This is somewhat fine as long as we know which timezone we're in. But for data storage and further data processing it would be great to store the start/end times as actual instants in time.

So this is what we do now! :)
We're using java.time.ZonedDateTime. It is indeed an unambiguous instant int time and it has nice human readable timezone info in the string that gets saved in the database.

#16