Fix: activity list not update on category update
This solution uses a LiveData<Unit> trigger that triggers a livedata switchMap that gets the activity list as livedata from the repository.

haveCategoriesChanged is a LiveData of Unit. We could have used the original category list as the value of the livedata, but using Unit gives me less chances to make bugs by accidentally using this category list here where it should not be used.

fixes #17