Refactor: update of activity list on category change
Connect the category table to the activity list query already in the SQL query in the DAO. Now Room takes care to update the category list livedata when either the categories or acitvities tables change. Neither repository no viewmodel know about this implementation.

These were my inspirations:
https://forums.asp.net/t/1060739.aspx?Selecting+Nothing+with+SQL+Statement
https://stackoverflow.com/questions/39766336/how-to-build-dummy-columns-for-union

The previous solution worked but had design drawbacks. ActivityListViewModel had to manually set up a second livedata that listens for category table changes and use that as a trigger to update the activity list.
1) A view model should not know or care about this "implementation" of the database layer.
2) We misuse livedata here for listening to database table changes. Because of A, such observations should be put in the repository, if at all, but livedata needs a viewLifecycleObserver which has no business in the repository.

#17