Make Activity check for data integrity and throw exception
Since Activity is an immutable data class, as long as we create valid, non-corrupted objects with the constructor (and all factory methods), the rest of the code can safely assume data integrity for existing objects.
So now the constructor (init block) checks that the end time is not before the start time and throws a custom ActivityTimeException otherwise.
Since the constructor doesn't know how to deal with this "corrupted" data, it simply throws this exception to the caller who will know how to respond to that.

#18