Refactor activity deletion dialog
It now uses a DialogFragment and ActivityListViewModel to handle screen rotations.

To share the view model, we have to use `by activityViewModels()` instead of `by viewModels()` because the latter gets a view model scoped to the current fragment, so it would create a new view model instance for each fragment, which is not what we want.

#22