Refactor ActivityAdapter: popup menu callbacks
Each activity list item has a popup menu, but what happens next is defined in the Fragment. So far we have 3 menu items and thus 3 callbacks, which were piped through from Fragment to ActivityAdapter to ViewHolder separately as lambda function parameters.

Now, this forwarding is done more elegantly in a separate data class holding the references to the functions.

This will be much more convenient to be extended in the future.

#10