Refactor ActivityListFragment
Limit the scope of the view binding from an instance variable to a local variable.

The subtlety with a view binding instance variable is that the fragment has a larger scope than its view, i.e. the view binding. In particular, the view is only there between onCreateView() and onDestroyView().

ActivityListFragment doesn't have many functions that use the view binding, so I can leave it as a local variable without having to worry about its scope.

#10