Introduce bottom navigation
The navigation drawer is fancy but not very user friendly. Users that don't know it, can't discover it easily as the icon is unsuspecting and wiping from left screen side is not something users do out of habit. (Also, the new back navigation by wiping from either side collides with this input.)

Bottom navigation is a lot more user friendly. It does take up space on the screen, but the discoverability may be worth it.

#13