Refactor category list: remove callbacks
Now that the CategoryListViewModel is shared among the dialog fragments and contains most of the logic, we can get rid of the callbacks for the popup menu for the category list elements.

Now we have to simply forward the viewModel via CategoryAdapter to the ViewHolders, that now know what to do, instead of the CategoryListFragment.

#7