Refactor activity list screen: eliminate callbacks
The previous solution had callbacks defined in the fragment, that were piped through ActivityAdapter to the ViewHolders.

Now with use of the shared view model, we could put most logic into that and don't need the callbacks.
We do have to pipe the viewModel to the viewHolders, though.

ActivityListFragment has become a lot smaller, cleaner, and dumber.

ActivityListViewModel took up some logic, but that logic is now decoupled from views or android context, so we can unit test it more easily. ^-^

#10